Rework handling of events (#89)
Changes how event attendance is tracked from a link sent during the
event to the link to the event itself. For example, last year organizers
would send a link in the meeting chat for participants to be marked as
having attended it. This year, when clicking on the link to join the
event, they are automatically marked as having attended, and then they
will be redirected to the meeting.

Furthermore, check-in is now a distinct process from an event with its
own link. This allows us to track registered participants vs attending
participants. Marking a participant as checked-in is also exposed to
WaffleBot, allowing us to determine attending participants from events
within Discord.

Modifications and deletions of events are published to NATS now under
the `workshops.automated.updated` and `workshops.automated.deleted`
topics for consumption by other internal services. This will be most
useful for syncing events to [Discord scheduled
events](https://support.discord.com/hc/en-us/articles/4409494125719-Scheduled-Events).